Split and modify the entry point
