+++ pkg/spanner-truncate/main.go
@@ -14,7 +14,6 @@
 // limitations under the License.
 //
 
-// spanner-truncate is a tool to delete all rows from the tables in a Cloud Spanner database without deleting tables themselves.
 package truncate
 
 import (
@@ -23,57 +22,11 @@ import (
 	"fmt"
 	"io"
 	"os"
-	"os/signal"
-	"strings"
 	"time"
 
 	"cloud.google.com/go/spanner"
 	"github.com/gosuri/uiprogress"
-	"github.com/jessevdk/go-flags"
 )
-
-type options struct {
-	ProjectID     string `short:"p" long:"project" env:"SPANNER_PROJECT_ID" description:"(required) GCP Project ID."`
-	InstanceID    string `short:"i" long:"instance" env:"SPANNER_INSTANCE_ID" description:"(required) Cloud Spanner Instance ID."`
-	DatabaseID    string `short:"d" long:"database" env:"SPANNER_DATABASE_ID" description:"(required) Cloud Spanner Database ID."`
-	Quiet         bool   `short:"q" long:"quiet" description:"Disable all interactive prompts."`
-	Tables        string `short:"t" long:"tables" description:"Comma separated table names to be truncated. Default to truncate all tables if not specified."`
-	ExcludeTables string `short:"e" long:"exclude-tables" description:"Comma separated table names to be exempted from truncating. 'tables' and 'exclude-tables' cannot co-exist"`
-}
-
-const maxTimeout = time.Hour * 24
-
-func main() {
-	var opts options
-	if _, err := flags.Parse(&opts); err != nil {
-		exitf("Invalid options\n")
-	}
-
-	if opts.ProjectID == "" || opts.InstanceID == "" || opts.DatabaseID == "" {
-		exitf("Missing options: -p, -i, -d are required.\n")
-	}
-
-	var targetTables []string
-	var excludeTables []string
-	if opts.Tables != "" {
-		targetTables = strings.Split(opts.Tables, ",")
-	}
-
-	if opts.ExcludeTables != "" {
-		if opts.Tables != "" {
-			exitf("Conflict: --tables and --exclude-tables cannot be both set.\n")
-		}
-		excludeTables = strings.Split(opts.ExcludeTables, ",")
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), maxTimeout)
-	defer cancel()
-	go handleInterrupt(cancel)
-
-	if err := Run(ctx, opts.ProjectID, opts.InstanceID, opts.DatabaseID, opts.Quiet, os.Stdout, targetTables, excludeTables); err != nil {
-		exitf("ERROR: %s", err.Error())
-	}
-}
 
 func Run(ctx context.Context, projectID, instanceID, databaseID string, quiet bool, out io.Writer, targetTables, excludeTables []string) error {
 	database := fmt.Sprintf("projects/%s/instances/%s/databases/%s", projectID, instanceID, databaseID)
@@ -137,11 +90,6 @@ func Run(ctx context.Context, projectID, instanceID, databaseID string, quiet bo
 
 	fmt.Fprint(out, "\nDone! All rows have been deleted successfully.\n")
 	return nil
-}
-
-func exitf(format string, a ...interface{}) {
-	fmt.Fprintf(os.Stderr, format, a...)
-	os.Exit(1)
 }
 
 // confirm returns true if a user confirmed the message, otherwise returns false.
@@ -213,11 +161,4 @@ func showProgressBar(progress *uiprogress.Progress, table *table, maxNameLength 
 			time.Sleep(time.Second * 1)
 		}
 	}()
-}
-
-func handleInterrupt(cancel context.CancelFunc) {
-	c := make(chan os.Signal, 1)
-	signal.Notify(c, os.Interrupt)
-	<-c
-	cancel()
 }
